No duplicate shared roster additions
Do not add users if they are already in the shared roster. In theory,
the operation is idempotent, but it seems that the roster change event
is still sent to the logged in users, resulting in them seeing the
entries twice until they log out and in again (at least on Gajim).